eldar-challenge-rodriguez/src/app/components/shared/nav-bar/nav-bar.component.ts
import { Component, inject } from '@angular/core';
import { ButtonModule } from 'primeng/button';
import { MenubarModule } from 'primeng/menubar';
import { AuthService } from '../../../services/auth.service';


@Component({
  selector: 'app-nav-bar',
  standalone: true,
  imports: [MenubarModule, ButtonModule],
  templateUrl: './nav-bar.component.html',
  styleUrl: './nav-bar.component.css'
})
export class NavBarComponent {

  authService = inject(AuthService)

  onLogoutClick() {
    this.authService.logout();
  }

}
